backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
sqlalchemy==2.0.36
alembic==1.18.4
asyncpg==0.30.0
psycopg2-binary==2.9.10
google-generativeai==0.8.4
qdrant-client==1.12.1
sentence-transformers==3.3.1
langchain-text-splitters==0.3.4
numpy==2.2.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1+cpu
